+++ src/app/core/services/gen-id.generator.ts
@@ -1,0 +1,10 @@
+export function* genID() {
+
+  let id = 1;
+  let nextId = 1;
+
+  while (true) {
+    yield [id, nextId] = [nextId, id + nextId];
+  }
+
+}