+++ src/app/shared/shared.module.ts
@@ -1,18 +1,21 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HighlightDirective } from './highlight.directive';
+import { FontResizeDirective } from './font-resize.directive';
 
 
 
 @NgModule({
   declarations: [
-    HighlightDirective
+    HighlightDirective,
+    FontResizeDirective,
   ],
   imports: [
     CommonModule
   ],
   exports: [
-    HighlightDirective
+    HighlightDirective,
+    FontResizeDirective,
   ]
 })
 export class SharedModule { }